+++ docs/CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.8.1:
 - The "unload" window event has been replaced with "pagehide" (due to "unload" being deprecated).
+- The public function "setConfiguration()" will now only update the configuration if something in the configuration has been changed.
 
 <br>